third_party/blink/web_tests/platform/generic/fast/events/menu-key-context-menu-position-expected.txt

Verifies that when element is focused in frames then selection by menu key, context menu should come on focused element. Succeeds if context menu is coming on the focused element and event co-ordinates are at center of the elements bounding box.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS event.target is inputNode
PASS event.clientX is X
PASS event.clientY is Y
